backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
requests
pandas
numpy
pytest
gunicorn
psycopg2-binary
python-dotenv
werkzeug
bcrypt
